#1183 - Support transaction profiling, move away from binary encoding

+++ src/main/java/io/ebeaninternal/api/ScopedTransaction.java
@@ -81,7 +81,7 @@ public class ScopedTransaction implements SpiTransaction {
   }
 
   @Override
-  public int profileOffset() {
+  public long profileOffset() {
     return transaction.profileOffset();
   }
