+++ src/main/java/io/ebeaninternal/server/core/PersistRequestUpdateSql.java
@@ -39,7 +39,7 @@ public final class PersistRequestUpdateSql extends PersistRequest {
   }
 
   @Override
-  public void profile(int offset, int flushCount) {
+  public void profile(long offset, int flushCount) {
     profileBase(EVT_UPDATESQL, offset, (short)0, flushCount);
   }
